+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## [Unreleased]
-### Changed
-- `auth.controller.js` `login()` no longer throws duplicate errors for incorrect username or password.
-- Remove `app.use(winstonInstance.errorLogger())` and replace with err handling middlware that calls `winstonInstance.info(err)`
+### Fixed
+- `auth.controller.js:login()` no longer throws duplicate errors for incorrect username or password
+- Use error handling middlware `winstonInstance.info(err)`
+  * Previously, `winstonInstance.errorLogger()` always threw a `TypeError`, meaning the underlying error didn't actually show in the logs
+
 
 ## [2.7.0] -- 2019-02-04
 ### Added