+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - `auth.controller.js:login()` no longer throws duplicate errors for incorrect username or password
 - Use error handling middlware `winstonInstance.info(err)`
   * Previously, `winstonInstance.errorLogger()` always threw a `TypeError`, meaning the underlying error didn't actually show in the logs
-
+- APIError now supports all 10 features in https://jira.amida.com/browse/SER-279
 
 ## [2.7.0] -- 2019-02-04
 ### Added